Determine whether a binding is expressible without Web IDL bindings
In the context of a JS embedder that does *not* implement the Web IDL
bindings proposal, are this binding's ingoing and outgoing conversions
losslessly expressible with the default conversions of the WebAssembly
JavaScript interface and the ECMAScript bindings for Web IDL?

This is primarily a question that is only relevant to polyfills for Web
IDL bindings, such as `wasm-bindgen`.

For both incoming and outgoing values, the conversions are not
expressible if each value is referenced more than once in the binding
map, or the values are referenced out of order. Failure to meet this
criteria would result in swapped, garbage, extra, or not enough
parameters/results.

In addition to the arity and usage requirements, each value's type must
also be trivially convertible. This property depends on whether the
value is incoming or outgoing.

When passing outgoing Wasm values to Web IDL, first they are converted
to JS values via the [`ToJSValue` algorithm defined by the WebAssembly
JavaScript
interface](https://webassembly.github.io/spec/js-api/index.html#tojsvalue). Next,
these JS values then they are converted into Web IDL values via [Web
IDL's ECMAScript type
mapping](https://heycam.github.io/webidl/#es-type-mapping).

```text
+------------+                                 +----------+                                        +---------------+
| Wasm value | ---WebAssembly-JS-interface---> | JS value | ---ECMAScript-bindings-for-Web-IDL---> | Web IDL value |
+------------+                                 +----------+                                        +---------------+
```

For outgoing values to be expressible without Web IDL bindings, nor
extra glue or conversions, they must be an `as` operator performing on
of the following conversions from a Wasm value type to a Web IDL type:

| From Wasm valtype             | To Web IDL Type     |
|-------------------------------|---------------------|
| `i32`, `f32`, `f64`, `anyref` | `any`                                                                               |
| `i32`                         | `long`, `long long`, `float`, `unrestricted float`, `double`, `unrestricted double` |
| `f32`                   | `unrestricted float`, `unrestricted double`                                               |
| `f64`                   | `unrestricted double`                                                                     |

When passing Web IDL values to Wasm, first they are converted to JS
values via the [Web IDL's ECMAScript type
mapping](https://heycam.github.io/webidl/#es-type-mapping), and then
those JS values are converted into Wasm values according to the
[`ToWebAssemblyValue` algorithm defined by the WebAssembly JavaScript
interface](https://webassembly.github.io/spec/js-api/index.html#towebassemblyvalue).

```text
+---------------+                                         +----------+                                 +---------------+
| Web IDL value |  ---ECMAScript-bindings-for-Web-IDL---> | JS value | ---WebAssembly-JS-interface---> | Web IDL value |
+---------------+                                         +----------+                                 +---------------+
```

For incoming values to be expressible without Web IDL bindings, nor
extra glue or conversions, they must be an incoming binding expression
of the form `(as (get <i>))` that is performing one of the following
conversions from a Web IDL type to a Wasm type:

| From Web IDL Type | To Wasm valtype |
|-------------------|-----------------|
| `any`                                                                                                                               | `anyref`        |
| `byte`, `short`, `long`                                                                                                             | `i32`           |
| `byte`, `octet`, `short`, `unsigned short`, `float`, `unrestricted float`                                                           | `f32`           |
| `byte`, `octet`, `short`, `unsigned short`, `long`, `unsigned long`, `float`, `unrestricted float`, `double`, `unrestricted double` | `f64`           |