Add a version marker header to binary encoding
This commit adds a version marker header to the temporary binary
encoding of the webidl custom section. The intention of this marker is
to help indicate to users when there's a mismatch between the encoder of
the webidl bindings section and the decoder. The section has no official
binary encoding and the spec is in a large state of flux right now, so
in this transitionary period there's likely to be a lot of development
and a lot of different tools which need to sync up. This version marker
is intended to assist with this by generating errors early on to
indicate that the versions are mismatched instead of generating odd
binary decoder errors.

The current version is just the version of this crate, but we can
probably use a fancier scheme in the future where we actually
deliberately bump the number. Presumably as soon as we coordinate with
other producers/consumers the scheme of using our own version number
will become obsolete. Hopefully by around that time we'll also be moving
towards an official spec encoding!